Fix memory leaks in main, disemvowel, and disemvowel_test.
All the leaks were freed in main because disemvowel() takes place in a while loop, and therefore doesn't know when it can free the memory it's being given.
So the line that was passed to it ended up being freed after the loop had taken place, and the memory returned by disemvowel(line) was stored in disemvoweled
in each iteration of the loop and freed after we had printed said return.

This didn't jive at all with the tests, which seemed to expect that disemvowel() would free it's returned memory on its own, which was obviously not the case in our implementation.
To fix this, we capture the pointer to the returned memory from disemvowel(), and free it after the ASSERT_STREQ statement to prevent memory leaks in the test.

As an aside, while our current implemenation works, our disemvowel() function should definitely have a helper function count_vowels() [that does what you'd think] that then passes that
info to disemvowel() [which would then be taking two inputs: an int and a pointer]. It's definitely two functions smushed together into one right now. If we end up having extra time,
we will probably refactor it.